xml/parsing/manakai/eldecls-perefs-1.dat

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!ELEMENT %fuga;EMPTY>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT xy EMPTY>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!ELEMENT %fuga; EMPTY>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT xy EMPTY>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (xy)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!ELEMENT ab (%fuga; )>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (xy)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!ELEMENT ab ( %fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (xy)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga " xy">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (xy)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy,aa">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (xy, aa)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy |  aa">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (xy | aa)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "|">
<!ELEMENT ab (bb%fuga;w)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (bb | w)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "+">
<!ELEMENT ab (foo%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;18;after element name;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "+">
<!ELEMENT ab foo%fuga;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;18;string after md def;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "(xy |  aa)">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab ((xy | aa))>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "(#PCDATA) ">
<!ELEMENT ab %fuga;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (#PCDATA)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "#PCDATA">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (#PCDATA)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "PCDATA">
<!ELEMENT ab (#%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;18;after element name;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "#PCDATA|abc">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (#PCDATA | abc)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "ab">
<!ELEMENT ab (%fuga;cd)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]2;21;after element name;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "ab">
<!ELEMENT ab (cd%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;18;after element name;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "DATA">
<!ELEMENT ab (#PC%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;18;after element name;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "TY">
<!ELEMENT ab EMP%fuga;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;18;string after md def;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "EMP">
<!ELEMENT ab %fuga;TY>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]2;20;string after md def;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "EMPTY>">
<!ELEMENT ab %fuga;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;23;mdc in pe in md;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "EMPTY>">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;23;mdc in pe in md;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
